+++ src/main/java/com/example/businessmanagement2/restcontroller/user/UserForm.java
@@ -1,22 +1,23 @@
 package com.example.businessmanagement2.restcontroller.user;
 
+
 import javax.validation.constraints.Size;
 import lombok.AllArgsConstructor;
-import lombok.Getter;
-
+import lombok.Data;
 import javax.validation.constraints.NotBlank;
 
 
-@Getter
+
+@Data
 @AllArgsConstructor
 public class UserForm {
 
+    @NotBlank
+    @Size(min = 1, max = 256)
+    private String companyname;
 
     @NotBlank
     @Size(min = 1, max = 256)
-    String companyname;
+    private String username;
 
-    @NotBlank
-    @Size(min = 1, max = 256)
-    String username;
 }
